+++ packages/playground/src/views/mc-plus/mc-table-plus/mock/apis.ts
@@ -1,0 +1,11 @@
+import { delay } from "@mc-plus/utils";
+import { POSITION_DATA } from "./constant";
+
+const getPositionList = async () => {
+  await delay(300);
+  return {
+    list: POSITION_DATA,
+  };
+};
+
+export { getPositionList };
