+++ packages/playground/src/views/mc-plus/mc-table-plus/mock/constant.ts
@@ -1,0 +1,56 @@
+const POSITION_DATA = [
+  {
+    tokenName: "xAAPL",
+    tokenDesc: "Apple Inc",
+    position: "10000",
+    unrealizedPnl: "353.43",
+    realizedPnl: "353.43",
+    todayPnl: "353.43",
+    isPositive: true,
+    changeRate: "0.25%",
+    price: "121.14",
+    quantity: "12421",
+    availableQuantity: "125252.21",
+  },
+  {
+    tokenName: "xTSLA",
+    tokenDesc: "Tesla Inc",
+    position: "10000",
+    unrealizedPnl: "1353.34",
+    realizedPnl: "1353.34",
+    isPositive: true,
+    changeRate: "0.25%",
+    todayPnl: "1353.34",
+    price: "121.14",
+    quantity: "12421",
+    availableQuantity: "125252.21",
+  },
+  {
+    tokenName: "xINTC",
+    tokenDesc: "Intel Corporation",
+    position: "10000",
+    unrealizedPnl: "353.43",
+    realizedPnl: "353.43",
+    isPositive: true,
+    changeRate: "0.25%",
+    todayPnl: "353.43",
+    price: "121.14",
+    quantity: "12421",
+    availableQuantity: "125252.21",
+  },
+  {
+    tokenName: "xLCID",
+    tokenDesc: "Lucid Group, Inc.",
+    position: "10000",
+    unrealizedPnl: "353.43",
+    realizedPnl: "353.43",
+    isPositive: true,
+    changeRate: "0.25%",
+    todayPnl: "353.43",
+    price: "121.14",
+    quantity: "12421",
+    availableQuantity: "125252.21",
+  },
+];
+
+export { POSITION_DATA };
