+++ packages/playground/src/views/mc-plus/mc-table-plus/tables/pagination-table/enums.ts
